requirements.txt
# Font Analysis and Generation Toolkit Requirements

# Core font processing
fonttools>=4.38.0
pillow>=9.3.0

# Numerical and scientific computing
numpy>=1.23.0
scipy>=1.9.0

# Machine Learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.64.0

# Jupyter support (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0

# TensorBoard for training visualization
tensorboard>=2.11.0

# scikit-learn for dimensionality reduction
scikit-learn>=1.2.0

# Additional utilities
pathlib>=1.0.0
requests>=2.28.0
